Project/Authentication System/djangojwt/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.5,<6.0
django-cors-headers>=4.8,<5.0
google-auth>=2.40,<3.0
